+++ crates/nanocld_client/src/lib.rs
@@ -2,6 +2,7 @@ mod http_client;
 
 pub(crate) mod namespace;
 pub(crate) mod cargo;
+pub(crate) mod exec;
 pub(crate) mod cargo_image;
 pub(crate) mod system;
 pub(crate) mod resource;
@@ -10,7 +11,7 @@ pub(crate) mod vm;
 pub(crate) mod vm_image;
 pub(crate) mod http_metric;
 pub(crate) mod node;
-
+pub use bollard_next;
 pub mod error;
 pub use http_client::*;
 pub use nanocl_stubs as stubs;