+++ bin/nanocl/changelog.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.10.0]
+
+### Added
+
+- Options for cargo exec: tty, detach_keys, env, privileged, user, working_dir  [@anonkey](https://github.com/anonkey)
+- return executed command status code from cargo exec [@anonkey](https://github.com/anonkey)
+
+### Changed
+
+- Use of nanocld-client v0.10.0 (exec_cargo becomes create_exec and start_exec) [@anonkey](https://github.com/anonkey)
+
+
 ## [0.9.0] - 2023-07-04
 
 ### Added